Move substitutions to context
