+++ src/preprocessing/assertion_pipeline.cpp
@@ -22,12 +22,7 @@
 namespace CVC4 {
 namespace preprocessing {
 
-AssertionPipeline::AssertionPipeline(context::Context* context)
-    : d_substitutionsIndex(context, 0),
-      d_topLevelSubstitutions(context),
-      d_realAssertionsEnd(0)
-{
-}
+AssertionPipeline::AssertionPipeline() : d_realAssertionsEnd(0) {}
 
 void AssertionPipeline::replace(size_t i, Node n)
 {
